feat: global commands
